NO-JIRA | feat: move handler to its own file
